Changes double to decimal

+++ src/Miningcore/Persistence/Postgres/Entities/Block.cs
@@ -5,7 +5,7 @@ public class Block
     public long Id { get; set; }
     public string PoolId { get; set; }
     public long BlockHeight { get; set; }
-    public double NetworkDifficulty { get; set; }
+    public decimal NetworkDifficulty { get; set; }
     public string Status { get; set; }
     public string Type { get; set; }
     public double ConfirmationProgress { get; set; }
