+++ src/Miningcore/Persistence/Postgres/Entities/PoolStats.cs
@@ -8,7 +8,7 @@ public class PoolStats
     public int ConnectedMiners { get; set; }
     public float PoolHashrate { get; set; }
     public double NetworkHashrate { get; set; }
-    public double NetworkDifficulty { get; set; }
+    public decimal NetworkDifficulty { get; set; }
     public DateTime? LastNetworkBlockTime { get; set; }
     public long BlockHeight { get; set; }
     public int ConnectedPeers { get; set; }
