+++ src/Miningcore/Persistence/Postgres/Entities/Share.cs
@@ -7,8 +7,8 @@ public class Share
     public string Miner { get; set; }
     public string Worker { get; set; }
     public string UserAgent { get; set; }
-    public double Difficulty { get; set; }
-    public double NetworkDifficulty { get; set; }
+    public decimal Difficulty { get; set; }
+    public decimal NetworkDifficulty { get; set; }
     public string IpAddress { get; set; }
     public string Source { get; set; }
     public DateTime Created { get; set; }
